+++ ProjetoVidaTOBE_Server/tobeApi/Controllers/InstitutionController.cs
@@ -2,19 +2,19 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
-using tobeApi.Data.Dtos.Scholling;
+using tobeApi.Data.Dtos.Institutions;
 using tobeApi.Models;
 using tobeApi.Services;
 
 namespace tobeApi.Controllers
 {
-    [Route("scholling")]
+    [Route("institution")]
     [ApiController]
-    public class SchollingController : ControllerBase
+    public class InstitutionController : ControllerBase
     {
-        private readonly SchollingService _service;
+        private readonly InstitutionService _service;
 
-        public SchollingController(SchollingService service)
+        public InstitutionController(InstitutionService service)
         {
             this._service = service;
         }
@@ -23,41 +23,49 @@ namespace tobeApi.Controllers
         [HttpGet]
         public IActionResult GetAll()
         {
-            IEnumerable<Scholling> scholling = _service.GetAll();
-            if (scholling == null) return NotFound();
-            return Ok(scholling);
+            IEnumerable<Institution> institution = _service.GetAll();
+            if (institution == null) return NotFound();
+            return Ok(institution);
         }
 
         [Authorize]
         [HttpGet("{id}")]
         public IActionResult Get(long id)
         {
-            Scholling scholling = _service.Get(id);
-            if (scholling == null) return NotFound();
-            return Ok(scholling);
+            Institution institution = _service.Get(id);
+            if (institution == null) return NotFound();
+            return Ok(institution);
         }
         [HttpPost]
         [Authorize(Roles = "admin")]
-        public IActionResult Create([FromBody] SchollingDto schollingDto)
+        public IActionResult Create([FromBody] InstitutionDto institutionDto)
         {
-            Scholling scholling = _service.Create(schollingDto);
-            if (scholling == null) return BadRequest();
-            return Ok(scholling);
+            Institution institution = _service.Create(institutionDto);
+            if (institution == null) return BadRequest();
+            return Ok(institution);
         }
 
         [HttpPut("{id}")]
         [Authorize(Roles = "admin")]
-        public IActionResult Update([FromBody] SchollingDto schollingDto, long id)
+        public IActionResult Update([FromBody] InstitutionDto institutionDto, long id)
         {
-            Scholling scholling = _service.Update(schollingDto, id);
-            if (scholling == null) return BadRequest();
-            return Ok(scholling);
+            if (_service.Get(id) == null)
+            {
+                return NotFound();
+            }
+            Institution institution = _service.Update(institutionDto, id);
+            if (institution == null) return BadRequest();
+            return Ok(institution);
         }
 
         [HttpDelete("{id}")]
         [Authorize(Roles = "admin")]
         public IActionResult Delete(long id)
         {
+            if (_service.Get(id) == null)
+            {
+                return NotFound();
+            }
             Result result = _service.Delete(id);
             if (result.IsFailed) return BadRequest();
             return Ok();
